Preprocess tag (locate and remove anchors)
