+++ reader.html
@@ -12,6 +12,7 @@
     <link rel="stylesheet" href="style/main.css">
     <link rel="stylesheet" href="style/github-corners.css"/>
     <script type="module" src="script/read/processSquare.js"></script>
+    <script type="module" src="script/read/preprocessSquare.js"></script>
     <script type="module" src="script/read/detectSquare.js"></script>
 </head>
 <body>
